pyproject.toml
[project]
name = "eor-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "google-genai (>=1.29.0,<2.0.0)",
    "chromadb (>=1.0.16,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
